+++ sklearn/utils/sparsefuncs.py
@@ -345,15 +345,41 @@ def count_nonzero(X, axis=None, sample_weight=None):
         raise ValueError('Unsupported axis: {0}'.format(axis))
 
 
-def csc_row_median(csc):
+def _get_median(data, n_zeros):
+    """Compute the median of data with n_zeros additional zeros.
+
+    This function is used to support sparse matrices; it modifies data in-place
     """
-    Find the median across axis 0 of a CSC matrix.
-    Wrapper aound the _get_median function in imputer and is equivalent
-    to doing np.median(X, axis=0)
+    n_elems = len(data) + n_zeros
+    if not n_elems:
+        return np.nan
+    n_negative = np.count_nonzero(data < 0)
+    middle, is_odd = divmod(n_elems, 2)
+    data.sort()
+
+    if is_odd:
+        return _get_elem_at_rank(middle, data, n_negative, n_zeros)
+
+    return (_get_elem_at_rank(middle - 1, data, n_negative, n_zeros) +
+            _get_elem_at_rank(middle, data, n_negative, n_zeros)) / 2.
+
+
+def _get_elem_at_rank(rank, data, n_negative, n_zeros):
+    """Find the value in data augmented with n_zeros for the given rank"""
+    if rank < n_negative:
+        return data[rank]
+    if rank - n_negative < n_zeros:
+        return 0
+    return data[rank - n_zeros]
+
+
+def csc_row_median(X):
+    """Find the median across axis 0 of a CSC matrix.
+    It is equivalent to doing np.median(X, axis=0).
 
     Parameters
     ----------
-    csc : CSC sparse matrix, shape (n_samples, n_features)
+    X : CSC sparse matrix, shape (n_samples, n_features)
         Input data.
 
     Returns
@@ -362,19 +388,17 @@ def csc_row_median(csc):
         Median. 
 
     """
-    from ..preprocessing.imputation import _get_median
+    if not isinstance(X, sp.csc_matrix):
+        raise TypeError("Expected matrix of CSC format, got %s" % X.format)
 
-    if not isinstance(csc, sp.csc_matrix):
-        raise TypeError("Expected matrix of CSC format, got %s" % csc.format)
-
-    indptr = csc.indptr
-    n_samples, n_features = csc.shape
+    indptr = X.indptr
+    n_samples, n_features = X.shape
     median = np.zeros(n_features)
 
     for f_ind, ptr in enumerate(indptr[:-1]):
 
-        # Prevent modifying csc in place
-        data = np.copy(csc.data[ptr: indptr[f_ind + 1]])
+        # Prevent modifying X in place
+        data = np.copy(X.data[ptr: indptr[f_ind + 1]])
         nz = n_samples - data.size
         median[f_ind] = _get_median(data, nz)
